feat: build form w/ formik
originally i wanted to build the form with react-core comps/props. but
that ended up with so much boilerplate that it would take away from the
goal of the project, which is to be able to compare frameworks. since
most react apps aren't going to do the boilerplate themselves it would
prove to be distracting.

the downside is that now the form is more of a demo of formik than
react. tradeoffs.